Animation builder.

+++ divoom/src/divoom_contracts/pixoo/animation/api_send_image_animation_frame.rs
@@ -32,7 +32,7 @@ impl DivoomPixooCommandAnimationSendImageAnimationFrameRequestPayload {
                 |entry| DivoomPixooCommandAnimationSendImageAnimationFrameRequestPayload {
                     pic_num: animation.frame_count as i32,
                     pic_width: animation.size as i32,
-                    pic_offset: entry.0,
+                    pic_offset: entry.0 as i32,
                     pic_id: animation.id,
                     pic_speed: animation.speed_in_ms,
                     pic_data: entry.1,
